requirements.txt
numpy>=1.26.4
pandas>=2.2.2
scipy>=1.13.0
matplotlib>=3.8.4
oclock==1.2.2
pyirsdk==1.3.5